Player Shotgun
Finally, the player has a shotgun. The spread is really poorly done (it's just a bunch of manually and haphazardly rotataed gameobjects), but it does the job.

Also resulted in a lot of scrap code from trying to figure out how to do it correctly, but all I learned is that rotations suck.